+++ src/components/containers/index.css
@@ -2,6 +2,6 @@
 
 @layer components {
     .grid3x2{
-        @apply grid xl:grid-cols-3 xl:grid-rows-2 grid-cols-1 grid-rows-6 xl:gap-6; 
+        @apply grid 2xl:grid-cols-3 2xl:grid-rows-2 grid-cols-1 grid-rows-6 2xl:gap-6; 
     }
 }
